pyproject.toml
[project]
name = "benchmark-360"
version = "1.0.0"
description = "Automatización de inteligencia competitiva para ISP en Ecuador"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "pydantic>=2.5.0",
    "lxml>=4.9.0",
    "aiohttp>=3.9.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.73.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
